+++ src/App.tsx
@@ -1,7 +1,7 @@
-import React, {FC} from "react"
-import {ThemeProvider} from "@mui/material"
-import {theme} from "theme"
-import {BrowserRouter, Route, Routes} from "react-router-dom"
+import React, { FC } from "react"
+import { ThemeProvider } from "@mui/material"
+import { theme } from "theme"
+import { BrowserRouter, Route, Routes } from "react-router-dom"
 import Signup from "pages/Signup"
 import Signin from "pages/Signin"
 import HomePage from "pages/MainScreen"
@@ -12,11 +12,11 @@ const App: FC = () => {
 		<ThemeProvider theme={theme}>
 			<BrowserRouter>
 				<Routes>
-					<Route path={"/sign-up"} element={<Signup/>}/>
+					<Route path={"/sign-up"} element={<Signup />} />
 
-					<Route path={"/sign-in"} element={<Signin/>}/>
-					<Route path={"/redirecting"} element={<Redirecting/>}/>
-					<Route path={"/"} element={<HomePage/>}/>
+					<Route path={"/sign-in"} element={<Signin />} />
+					<Route path={"/redirecting"} element={<Redirecting />} />
+					<Route path={"/"} element={<HomePage />} />
 				</Routes>
 			</BrowserRouter>
 		</ThemeProvider>
