+++ src/elements/button.ts
@@ -5,7 +5,7 @@ export const RoundedButton = styled(Button)<ButtonProps>((theme) => ({
 	borderRadius: "5em",
 	padding: ".75rem 1.7rem",
 	fontSize: ".75rem",
-	margin:"0 auto"
+	margin: "0 auto",
 }))
 
 RoundedButton.defaultProps = {
