+++ src/index.css
@@ -10,6 +10,3 @@
 html {
 	font-size: 14px;
 }
-
-
-
