+++ src/layouts/AuthenticateLayout/styles.tsx
@@ -95,4 +95,3 @@ export const navigationTitle: CSSProperties = {
 	fontSize: "0.8rem",
 	transform: "translateY(4rem)",
 }
-
