+++ src/types/IAuth/index.ts
@@ -1,0 +1,9 @@
+export interface ICredentials {
+	email: string
+	password: string
+}
+
+export interface IValidation {
+	helperText: string
+	error: boolean
+}
