+++ src/utils/index.ts
@@ -1,0 +1,3 @@
+export function isValidEmail(email: string): boolean {
+	return /^\w+([.-]?\w+)*@\w+([.-]?\w+)*(\.\w{2,3})+$/.test(email)
+}
